streamlined behaviour and strafing
Targeting and menu navigation have been streamlined to behave similar. Targeting wont close automatically instead it uses a select or clear button. The menu can be cycled forward and backward. Buttons for navigating the menu/targeting do not conflict with each other or steering so cycling can always be done with e.g. triggers as well as strafing. This approach of disconnecting targeting, like the menu, means only one button during steering must be dedicated to opening targeting/menu. The additional buttons are used for strafing.

Added additional binding for targeting.

Added additional bindings for menu navigation.

Added strafing to the "directional" steering mode.

Fixed a bug in which the default value for "axis or buttons" segments was null while also forbidding null, which ment if the user didn't edit the entry it would invalidate the entire configuration.